move to inventory service
