+++ src/Snap.Hutao/Snap.Hutao/Service/Inventory/IInventoryDbService.cs
@@ -2,16 +2,21 @@
 // Licensed under the MIT license.
 
 using Snap.Hutao.Model.Entity;
+using Snap.Hutao.Service.Abstraction;
 
 namespace Snap.Hutao.Service.Inventory;
 
-internal interface IInventoryDbService
+internal interface IInventoryDbService : IAppDbService<InventoryItem>
 {
-    ValueTask AddInventoryItemRangeByProjectId(List<InventoryItem> items);
+    ValueTask AddInventoryItemRangeByProjectIdAsync(List<InventoryItem> items, CancellationToken token = default);
 
-    ValueTask RemoveInventoryItemRangeByProjectId(Guid projectId);
+    ValueTask RemoveInventoryItemRangeByProjectIdAsync(Guid projectId, CancellationToken token = default);
 
     void UpdateInventoryItem(InventoryItem item);
 
-    ValueTask UpdateInventoryItemAsync(InventoryItem item);
+    ValueTask UpdateInventoryItemAsync(InventoryItem item, CancellationToken token = default);
+
+    List<InventoryItem> GetInventoryItemListByProjectId(Guid projectId);
+
+    ValueTask<List<InventoryItem>> GetInventoryItemListByProjectIdAsync(Guid projectId, CancellationToken token = default);
 }
